R interface to yacas computer algebra package

Ryacas0 allows one to use the yacas computer algebra package entirely from within R. It takes an R expression, an R one line function or a yacas string and returns an R expression or a variety of other formats. It can be used for symbolic mathematics, exact arithmetic, ASCII pretty printing and R to TeX conversions. The main command is yacas and ?yacas provides some information on installation and startup.

Details

The following are sources of information on "Ryacas0":

DESCRIPTION file	library(help = Ryacas0)

List of demo files	demo(package = "Ryacas0")

Demo file	demo("Ryacas0")

Demo	demo("Ryacas0-PrettyPrinter")

Demo	demo("Ryacas0-Function")

Demo	demo("Ryacas0-Sym")

Demo	demo("Ryacas0-Expr")

List Vignettes	vignette(package = "Ryacas0")

Vignette	vignette("Ryacas0")

This File	package?Ryacas0

Help files	?yacas, ?yacasTranslations, ?yacmode, ?Sym

Help files - Windows	?yacasInstall

News	RShowDoc("NEWS", package = "Ryacas0")

Acknowledgements	RShowDoc("THANKS", package = "Ryacas0")

Wish List	RShowDoc("WISHLIST", package = "Ryacas0")

Home page	https://github.com/r-cas/ryacas0/

Note

There is a note in the help file of the yacas command that discusses a number of installation and startup issues.

Examples

print(yacas(expression(integrate(1/x, x))))
#> yacas_expression(log(x))
print(yacas("Integrate(x)1/x"))
#> yacas_expression(log(x))
x <- Sym("x"); Integrate(1/x, x)
#> yacas_expression(log(x))
acos(Sym("1/2"))
#> yacas_expression(pi/3)