Yacas translations

Translations from R to the yacas computer algebra system.

Note

The translation process occurs in several steps. If the input to the yacas function is an expression then it is translated to a valid yacas character string (otherwise, it is sent to yacas unprocessed). Yacas then processes the string and if retclass="expression" it is translated back to an R expression (otherwise it is sent back unprocessed). Currently supported translations are:

CONSTANTS

R		yacas

=		=====

pi		Pi

OPERATORS

R		yacas

=		=====

7 \%\% 3		Mod(7, 3)

7 \%\/\% 3		Div(7, 3)

FUNCTIONS

R		yacas

=		=====

sin(x)		Sin(x)

cos(x)		Cos(x)

tan(x)		Tan(x)

asin(x)		ArcSin(x)

acos(x)		ArcCos(x)

atan(x)		ArcTan(x)

exp(x)		Exp(x)

sqrt(x)		Sqrt(x)

log(x)		Ln(x)

choose(n, k)		Bin(n, k)

gamma(x)		Gamma(x)

deriv(sin, x)		Deriv(x)Sin(x)

integrate(f, a, b)		Integrate(x, a, b)f(x)

list()		List()

factorial(n)		n!

Note the Limit example in demo(Ryacas0) for adding translations on the fly.

The complete table under development.

References

http://www.yacas.org/

Author

Rob J Goedman